return 1 in docs

+++ docs.sh
@@ -6,21 +6,22 @@ then
   cd out
   git clone https://github.com/srora/cloudsuite.git
   cd cloudsuite
-  #git remote add upstream https://github.com/srora/cloudsuite.git
-  #git fetch upstream
-  #git checkout -b gh-pages --track upstream/gh-pages
   git checkout gh-pages
   git config user.name ${GIT_USER}
   git config user.email ${GIT_EMAIL}
   git config credential.helper "store --file=.git/credentials"
   git config --global push.default matching
   echo "https://$GH_TOKEN:x-oauth-basic@github.com" >> .git/credentials
-  #rm -rf docs
-  #cp -R ../../docs ./docs
-  #git add .
-  #git commit -a -m "Rebuilding"
   git merge master --no-edit
+  if [ $result -eq "1" ]
+  then
+      return 1
+  fi
   git push -f origin gh-pages
+  if [ $result -eq "1" ]
+  then
+      return 1
+  fi
   cd ${TRAVIS_BUILD_DIR}
   rm -rf out
   git checkout master
